Move ee code to ee package and unify ee status to IsPro
