+++ servercfg/serverconf.go
@@ -19,7 +19,7 @@ const EmqxBrokerType = "emqx"
 
 var (
 	Version              = "dev"
-	Is_EE                = false
+	IsPro                = false
 	ErrLicenseValidation error
 )
 
@@ -88,7 +88,7 @@ func GetServerConfig() config.ServerConfig {
 	cfg.StunList = GetStunListString()
 	cfg.Verbosity = GetVerbosity()
 	cfg.IsEE = "no"
-	if Is_EE {
+	if IsPro {
 		cfg.IsEE = "yes"
 	}
 
@@ -110,7 +110,7 @@ func GetServerInfo() models.ServerConfig {
 		cfg.DNSMode = "on"
 	}
 	cfg.Version = GetVersion()
-	cfg.Is_EE = Is_EE
+	cfg.Is_EE = IsPro
 	cfg.StunPort = GetStunPort()
 	cfg.StunList = GetStunList()
 	cfg.TurnDomain = GetTurnHost()
